Module type Expect_test_config_types.S

module IO : sig ... end

val run : (unit -> unit IO.t) -> unit

Run an IO operation until completion

val sanitize : string -> string

sanitize can be used to map all output strings, e.g. for cleansing.

val upon_unreleasable_issue : Upon_unreleasable_issue.t

upon_unreleasable_issue specifies how to deal with output that should not be released even if it is accepted (e.g. backtraces). The default is `CR.